+++ pkg/actors/router/router.go
@@ -110,16 +110,17 @@ func (r *router) Call(ctx context.Context, req *internalv1pb.InternalInvokeReque
 	var res *internalv1pb.InternalInvokeResponse
 	if r.resiliency.PolicyDefined(req.GetActor().GetActorType(), resiliency.ActorPolicy{}) {
 		res, err = r.callActor(ctx, req)
-		// Don't bubble perminant errors up to the caller to interfere with top level
-		// retries.
-		if _, ok := err.(*backoff.PermanentError); ok {
-			err = errors.Unwrap(err)
-		}
 	} else {
 		policyRunner := resiliency.NewRunner[*internalv1pb.InternalInvokeResponse](ctx, r.resiliency.BuiltInPolicy(resiliency.BuiltInActorNotFoundRetries))
 		res, err = policyRunner(func(ctx context.Context) (*internalv1pb.InternalInvokeResponse, error) {
 			return r.callActor(ctx, req)
 		})
+	}
+
+	// Don't bubble perminant errors up to the caller to interfere with top level
+	// retries.
+	if _, ok := err.(*backoff.PermanentError); ok {
+		err = errors.Unwrap(err)
 	}
 
 	return res, err
