+++ pkg/actors/targets/workflow/sink.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2024 The Dapr Authors
+Copyright 2025 The Dapr Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -13,9 +13,18 @@ limitations under the License.
 
 package workflow
 
-import (
-	_ "github.com/dapr/dapr/tests/integration/suite/daprd/workflow/continueasnew"
-	_ "github.com/dapr/dapr/tests/integration/suite/daprd/workflow/memory"
-	_ "github.com/dapr/dapr/tests/integration/suite/daprd/workflow/rerun"
-	_ "github.com/dapr/dapr/tests/integration/suite/daprd/workflow/scheduler"
-)
+import "github.com/dapr/durabletask-go/backend"
+
+func (w *workflow) runEventSink(ch chan *backend.OrchestrationMetadata, cb func(*backend.OrchestrationMetadata)) {
+	for {
+		select {
+		case <-w.closeCh:
+			return
+		case val, ok := <-ch:
+			if !ok {
+				return
+			}
+			cb(val)
+		}
+	}
+}
