+++ tests/integration/suite/daprd/workflow/rerun/basic.go
@@ -1,0 +1,80 @@
+/*
+Copyright 2025 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://wwb.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package rerun
+
+import (
+	"context"
+	"sync/atomic"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/dapr/dapr/tests/integration/framework"
+	"github.com/dapr/dapr/tests/integration/framework/process/workflow"
+	"github.com/dapr/dapr/tests/integration/suite"
+	"github.com/dapr/durabletask-go/api"
+	"github.com/dapr/durabletask-go/task"
+)
+
+func init() {
+	suite.Register(new(basic))
+}
+
+type basic struct {
+	workflow *workflow.Workflow
+}
+
+func (b *basic) Setup(t *testing.T) []framework.Option {
+	b.workflow = workflow.New(t)
+
+	return []framework.Option{
+		framework.WithProcesses(b.workflow),
+	}
+}
+
+func (b *basic) Run(t *testing.T, ctx context.Context) {
+	b.workflow.WaitUntilRunning(t, ctx)
+
+	var act atomic.Int64
+	b.workflow.Registry().AddOrchestratorN("foo", func(ctx *task.OrchestrationContext) (any, error) {
+		for range 5 {
+			require.NoError(t, ctx.CallActivity("bar").Await(nil))
+		}
+		return nil, nil
+	})
+	b.workflow.Registry().AddActivityN("bar", func(ctx task.ActivityContext) (any, error) {
+		act.Add(1)
+		return nil, nil
+	})
+	client := b.workflow.BackendClient(t, ctx)
+
+	id, err := client.ScheduleNewOrchestration(ctx, "foo", api.WithInstanceID("abc"))
+	require.NoError(t, err)
+	_, err = client.WaitForOrchestrationCompletion(ctx, id)
+	require.NoError(t, err)
+	assert.Equal(t, int64(5), act.Load())
+
+	for i := range uint32(5) {
+		act.Store(0)
+
+		newID, err := client.RerunWorkflowFromEvent(ctx, api.InstanceID("abc"), i)
+		require.NoError(t, err)
+		assert.NotEqual(t, id, newID)
+
+		_, err = client.WaitForOrchestrationCompletion(ctx, newID)
+		require.NoError(t, err)
+		assert.Equal(t, int64(5-i), act.Load())
+	}
+}
