+++ tests/integration/suite/daprd/workflow/rerun/multirerun.go
@@ -1,0 +1,79 @@
+/*
+Copyright 2025 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://wwb.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package rerun
+
+import (
+	"context"
+	"sync/atomic"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"github.com/dapr/dapr/tests/integration/framework"
+	"github.com/dapr/dapr/tests/integration/framework/process/workflow"
+	"github.com/dapr/dapr/tests/integration/suite"
+	"github.com/dapr/durabletask-go/api"
+	"github.com/dapr/durabletask-go/task"
+)
+
+func init() {
+	suite.Register(new(multirerun))
+}
+
+type multirerun struct {
+	workflow *workflow.Workflow
+}
+
+func (m *multirerun) Setup(t *testing.T) []framework.Option {
+	m.workflow = workflow.New(t)
+
+	return []framework.Option{
+		framework.WithProcesses(m.workflow),
+	}
+}
+
+func (m *multirerun) Run(t *testing.T, ctx context.Context) {
+	m.workflow.WaitUntilRunning(t, ctx)
+
+	var act atomic.Int64
+	m.workflow.Registry().AddOrchestratorN("foo", func(ctx *task.OrchestrationContext) (any, error) {
+		require.NoError(t, ctx.CallActivity("bar").Await(nil))
+		return nil, nil
+	})
+	m.workflow.Registry().AddActivityN("bar", func(ctx task.ActivityContext) (any, error) {
+		act.Add(1)
+		return nil, nil
+	})
+
+	client := m.workflow.BackendClient(t, ctx)
+
+	id, err := client.ScheduleNewOrchestration(ctx, "foo", api.WithInstanceID("abc"))
+	require.NoError(t, err)
+	_, err = client.WaitForOrchestrationCompletion(ctx, id)
+	require.NoError(t, err)
+
+	assert.Equal(t, int64(1), act.Load())
+
+	newIDs := make([]api.InstanceID, 10)
+	for i := range 10 {
+		newIDs[i], err = client.RerunWorkflowFromEvent(ctx, api.InstanceID("abc"), 0)
+		require.NoError(t, err)
+	}
+	for i := range 10 {
+		_, err = client.WaitForOrchestrationCompletion(ctx, newIDs[i])
+		require.NoError(t, err)
+	}
+	assert.Equal(t, int64(11), act.Load())
+}
